Rework healpix response computations.
The faster_geometry experimental module replaces fast_geometry for healpix fixed skymaps.
It can be activated with the boolean flag "faster_geometry" in params.

The class submodel (as well as SubmodelSpec and SubmodelKind) is also factored out into
its own file to avoid circular imports.
