+++ tests/params_test_faster_geometry.ini
@@ -1,0 +1,57 @@
+[params]
+fmin=3e-4
+fmax=2e-3
+duration=3e5
+seglen=1e5
+fs=0.25
+nside=2
+lmax=1
+lisa_config=orbiting
+tdi_lev=xyz
+load_data=no
+faster_geometry=yes
+
+model=noise+truncatedpowerlaw_fixedbulge+truncatedpowerlaw_fixeddisk
+
+alias = {"noise": "noise",
+         "truncatedpowerlaw_fixedbulge": "truncatedpowerlaw_fixedbulge",
+         "truncatedpowerlaw_fixeddisk": "truncatedpowerlaw_fixeddisk"}
+
+[inj]
+doInj=yes
+inj_only=no
+
+injection=noise+truncatedpowerlaw_fixedbulge+truncatedpowerlaw_fixeddisk
+
+truevals={
+    "truncatedpowerlaw_fixedbulge": {
+        'alpha': 0.82,
+        'log_omega0': -8.27,
+        'log_fcut': -2.785,
+        'log_fscale': -3.407,
+    },
+    "truncatedpowerlaw_fixeddisk": {
+        'alpha': 0.82,
+        'log_omega0': -8.90,
+        'log_fcut': -2.684,
+        'log_fscale': -3.82,
+    },
+    "noise": {
+        'Np': 9e-42,
+        'Na': 3.6e-49,
+    }}
+
+[run_params]
+out_dir=this_directory_should_not_exist/
+sampler=numpyro
+N_GPU=1
+Nthreads=1
+Nburn=100
+Nsamples=400
+show_progress=1
+checkpoint=false
+
+FixSeed=1
+seed=42
+
+projection=G
